requirements.txt
python-telegram-bot>=20.0
web3>=6.0.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
aiohttp>=3.8.0
matplotlib>=3.5.0
pandas>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter
asyncpg>=0.27.0  # Async PostgreSQL adapter
pydantic>=2.0.0
